+++ src/main/java/kitchenpos/menus/tobe/application/MenuService.java
@@ -3,9 +3,9 @@ package kitchenpos.menus.tobe.application;
 import kitchenpos.common.infra.Profanities;
 import kitchenpos.menugroups.domain.MenuGroup;
 import kitchenpos.menugroups.domain.MenuGroupInfo;
-import kitchenpos.menuproducts.domain.MenuProductTranslator;
-import kitchenpos.menuproducts.domain.MenuProducts;
 import kitchenpos.menus.tobe.domain.*;
+import kitchenpos.menus.tobe.domain.menuproducts.MenuProductTranslator;
+import kitchenpos.menus.tobe.domain.menuproducts.MenuProducts;
 import kitchenpos.menus.tobe.dto.ChangeMenuPriceRequest;
 import kitchenpos.menus.tobe.dto.CreateMenuRequest;
 import kitchenpos.menus.tobe.dto.MenuResponse;
