+++ src/main/java/kitchenpos/menus/tobe/dto/CreateMenuRequest.java
@@ -1,7 +1,5 @@
 package kitchenpos.menus.tobe.dto;
 
-import kitchenpos.menuproducts.dto.MenuProductRequest;
-
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
